pyproject.toml
[project]
name = "wisegpt"
version = "0.1.0"
description = "A CLI application that provides wise advice"
authors = [{name = "Ognjen Bostjancic"}]
dependencies = [
    "click>=8.0.0",
    "litellm>=1.30.7",
    "python-dotenv>=1.0.0",
    "sentry-sdk>=1.40.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.2.0",
    "ruff>=0.3.0",
    "pre-commit>=3.6.0",
]

[project.scripts]
wisegpt = "wisegpt.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py39"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = ["E501"]  # line length handled by black

[tool.ruff.isort]
known-first-party = ["wisegpt"]

[tool.black]
line-length = 88
target-version = ["py39"]
include = '\.pyi?$'
extend-exclude = '''
# A regex preceded with ^/ will apply only to files and directories
# in the root of the project.
^/tests/
'''
